requirements.txt
Keras==2.3.1
aiohttp==3.7.3
pyzmq==19.0.2
numpy==1.16.4
pandas==1.1.3
opencv-python==3.4.1.15
matplotlib==3.3.2
